Makefile: Fix detection of clang when cross-compiling
When the host's C compiler is clang, and when attempting to
cross-compile Linux e.g. to MIPS with mipsel-linux-gcc, the Makefile
would incorrectly detect the use of clang, which resulted in
clang-specific flags being passed to mipsel-linux-gcc.

This can be verified under Debian by installing the "clang" package,
and then using it as the default compiler with:
sudo update-alternatives --config cc

This patch moves the detection of clang after the $(CC) variable is
initialized to the name of the cross-compiler, so that the check applies
to the cross-compiler and not the host's C compiler.

v2: Move the detection of clang after the inclusion of the
arch/*/Makefile (as they might set $(CROSS_COMPILE))

Signed-off-by: Paul Cercueil <paul@crapouillou.net>
Signed-off-by: Michal Marek <mmarek@suse.cz>
(cherry picked from https://github.com/nathanchance/android-kernel-clang/commit/c0ebf280501bb84673e7c1eda00eca97e4914c18)
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Change-Id: I01bde823792eb691546d2b90a00649195d0b5634